Moved @Column to the annotation package
