+++ src/com/activeandroid/annotation/Computed.java
@@ -1,4 +1,4 @@
-package com.activeandroid;
+package com.activeandroid.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
